README.md
https://img.shields.io/github/issues/AmazingDaniel/Homework1
https://img.shields.io/github/forks/AmazingDaniel/Homework1
https://img.shields.io/github/stars/AmazingDaniel/Homework1
https://img.shields.io/github/license/AmazingDaniel/Homework1
https://img.shields.io/twitter/url?url=https%3A%2F%2Fgithub.com%2FAmazingDaniel%2FHomework1%2Fblob%2Fmaster%2Findex.html

Credit: https://zoom.us/rec/share/SVtp8_Rz6-u87unqlUQHyHIV4eYm3sdt5IREkVN_y-ZkeAGv9_2ZCAN8xP5W1kI.V5MeQJ1EGRHPEQl6?startTime=1600181993000

I started this project to change and add elements and missing words to make the website work. The end goal of the website was to make it accessible to the physical impaired and fix any code that wasn't complete. Move some of the code to clean it up.

MIT License

Copyright (c) [year] [fullname]

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.